refactor: move gArgs usages from llmq/utils to llmq/options
