+++ src/llmq/options.h
@@ -8,14 +8,37 @@
 #include <llmq/params.h>
 #include <gsl/pointers.h>
 
+#include <map>
+
 class CBlockIndex;
 
 namespace llmq
 {
+
+enum class QvvecSyncMode {
+    Invalid = -1,
+    Always = 0,
+    OnlyIfTypeMember = 1,
+};
+
+static constexpr bool DEFAULT_ENABLE_QUORUM_DATA_RECOVERY{true};
+
+// If true, we will connect to all new quorums and watch their communication
+static constexpr bool DEFAULT_WATCH_QUORUMS{false};
+
 bool IsAllMembersConnectedEnabled(Consensus::LLMQType llmqType);
 bool IsQuorumPoseEnabled(Consensus::LLMQType llmqType);
 
 bool IsQuorumRotationEnabled(const Consensus::LLMQParams& llmqParams, gsl::not_null<const CBlockIndex*> pindex);
+
+/// Returns the state of `-llmq-data-recovery`
+bool QuorumDataRecoveryEnabled();
+
+/// Returns the state of `-watchquorums`
+bool IsWatchQuorumsEnabled();
+
+/// Returns the parsed entries given by `-llmq-qvvec-sync`
+std::map<Consensus::LLMQType, QvvecSyncMode> GetEnabledQuorumVvecSyncEntries();
 
 } // namespace llmq
